Renamed Files for Better Consistency
+ All filenames now share the same naming convention.
